implement errata item #14
there has been the problem that occasionaly, an incoming package would
be entered into the ring buffer, but ENC_EPKTCNT would still show 0;
thus, the package is only read when the next package is received, which
itself ends up in the same position.

as a result, a typical ping timout was 1007ms (7ms normally, and 1s the
sender waited until the next ping was sent, which freed the first one).

also, the effect would stack up to several seconds of ping delay.

after this patch, it could not be triggered to more than 1s of delay. (a
flood ping reliably produced that problem).