drop set-read-pointer optimization
the value was not properly calculated, leading to the read pointer being
anywere and only garbage being received. the optimization did not give
much benefit anyway, because most of the time the read pointer still had
to be changed due to the alignment, and inside a package reception,
addresses are not set anymore anyway.