use the emlib as shipped in mbed
as siliconlabs didn't manage to make emlib available for download,
switching to using the copies shipped in mbed. (as an extra, this
contains an apache licensed copy of their usb library).

the library sources now indicated do not contain headers for the
specific hardware suggested in the README file, this is still
investigated.